chore: update useMouseHover.ts
